requirements.txt
streamlit>=1.28.0
redis>=5.0.0
python-dotenv>=1.0.0
PyYAML>=6.0
llama-index>=0.9.0
llama-index-vector-stores-chroma>=0.1.0
chromadb>=0.4.0
pdfplumber>=0.9.0
trafilatura>=1.6.0
lxml_html_clean>=0.1.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
requests>=2.31.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
sentence-transformers>=2.2.0
python-docx>=0.8.11